shuffle handlers out
